\name{albinia}
\alias{albinia}
\docType{data}
\title{
Céramiques romaines antiques de Albinia, Etrurie.
}
\description{
Dataset issu de la thèse de (Thierrin-Michae, 1992).
}
\usage{data("albinia")}
\format{
  A data frame with 25 observations on the following 24 variables.
  \describe{
    \item{\code{CaO}}{a numeric vector}
    \item{\code{Fe2O3}}{a numeric vector}
    \item{\code{TiO2}}{a numeric vector}
    \item{\code{K2O}}{a numeric vector}
    \item{\code{SiO2}}{a numeric vector}
    \item{\code{Al2O3}}{a numeric vector}
    \item{\code{MgO}}{a numeric vector}
    \item{\code{MnO}}{a numeric vector}
    \item{\code{Na2O}}{a numeric vector}
    \item{\code{P2O5}}{a numeric vector}
    \item{\code{Zr}}{a numeric vector}
    \item{\code{Sr}}{a numeric vector}
    \item{\code{Rb}}{a numeric vector}
    \item{\code{Zn}}{a numeric vector}
    \item{\code{Cr}}{a numeric vector}
    \item{\code{Ni}}{a numeric vector}
    \item{\code{La}}{a numeric vector}
    \item{\code{Ba}}{a numeric vector}
    \item{\code{V}}{a numeric vector}
    \item{\code{Ce}}{a numeric vector}
    \item{\code{Y}}{a logical vector}
    \item{\code{Th}}{a logical vector}
    \item{\code{Pb}}{a logical vector}
    \item{\code{Cu}}{a logical vector}
  }
}

\source{
Romische Weinamphoren,1992,Gisela Thierrin-Michael,These de l'Universite de Fribourg,Suisse
}
\examples{
data(albinia)
## maybe str(albinia) ; plot(albinia) ...
}
\keyword{datasets}
