% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_discriminante.R
\name{ana_dis}
\alias{ana_dis}
\title{Analyse discriminante quadratique}
\usage{
ana_dis(data, pas, echantillon, normalize = TRUE, echantillonage = TRUE)
}
\arguments{
\item{data}{Un dataframe. Les données sur lesquelles l'analyse sera effectuée.}

\item{pas}{Un sclaire. Il correspond à la largeur des intervalles de histogramme.}

\item{echantillon}{Un dataframe d'un ou plusieurs individus. Il s'agit des données pourlequelles il faut établir le groupe de référence.}

\item{normalize}{Un booleen. Si TRUE, les datas sont ramenées à 100 avant la classification, sinon elles sont gardées telles quelles.}

\item{echantillonage}{Un booleen. Si TRUE, le nom des échantillons apparaît sur l'histogramme.}
}
\value{
L'histogramme des distances euclidennes et de Mahalannobis.
}
\description{
L'analyse discriminante quadratique permet d'attribuer une céramique donnée à un groupe de référence à l'aide d'un calcul de probabilité d'appartenance. On identifie au préalable dans la base donnée un certain nombre de groupes parmi lesquels on cherche à identifier celui auquel appartient la céramique étudiée.
Cette attribution s'exprime en termes de statistique par la distance de Mahalannobis et un histogramme est tracé.
}
\details{
\cr La probabilité d'appartenance d'une céramique x au groupe s'exprime par :
\deqn{f_{k}(x) = \frac{e^{-\frac{1}{2}(x - \mu_{k})' C_{k}^{-1} (x - \mu_{k})}}{(2\pi)^{\frac{n}{2}} \left| C_{k} \right|^{\frac{1}{2}}}}

où :
\itemize{
\item \eqn{\mu_k} est la moyenne des éléments du groupe \eqn{k},
\item \eqn{C_k} est la matrice de covariance du groupe \eqn{k},
\item \eqn{n} est le nombre de variables.
}

Cette approche est néanmoins entièremment géométrique et ne prend pas en compte les probabilités \emph{a priori} des différents groupes.
}
\examples{

#data=as.dataframe(c(1,2,3))
#ana_dis(data)



}
\references{
Picon, M. (1984). \emph{PACT 10}. Academic Press. \cr
M. J. Baxter(1994), \emph{Exploratory multivariate analysis in archeology}, Editions Edinburgh University Press. \cr
Chenorkian, R. (1996), \emph{Pratique archéologique statistique et graphique}, Editions Errance \cr
}
