% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaire.R
\name{binaire}
\alias{binaire}
\title{Diagramme binaire}
\usage{
binaire(data, x, y, groupe = FALSE)
}
\arguments{
\item{data}{Un dataframe.}

\item{x}{Un string. Le nom de l'élément que l'on veut en abssice.}

\item{y}{Un string. Le nom de l'élément que l'on veut en ordonnée.}

\item{groupe}{Un booléen. Si TRUE, les points du biplot sont coloriés en fonction de la dernière colonne de \code{data}.}
}
\value{
Un biplot du diagramme binaire.
}
\description{
En analyse de données, en archéométrie, il est souvent utile de tracer des diagrammes binaires entre deux éléments des variables d'analyse. Cette fonction permet de tracer un biplot pour un dataset du laboratoire.
}
\examples{
## dataset de Waksman, 1999
data('montpellier')
binaire(montpellier,'CaO','Fe2O3')

}
