% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_de_grappe.R
\name{grappe}
\alias{grappe}
\title{L'analyse de grappe ou classification hiérarchique ascendante est une technique de classification classique en analyse de données. Pour une revue synthétique, voir Gordon 1987}
\usage{
grappe(data, normalize = TRUE, methode = "simple", h = NULL, k = NULL)
}
\arguments{
\item{data}{fichier data}

\item{normalize}{booleen. Vrai si on ramène à 100, faux sinon. Vrai par défaut}

\item{methode}{string. methode d'analyse simple ou log-lineaire}

\item{h}{hauteur decoupage groupe}

\item{k}{nombre decoupage groupe}
}
\value{
hc un dendogramme
}
\description{
L'analyse de grappe ou classification hiérarchique ascendante est une technique de classification classique en analyse de données. Pour une revue synthétique, voir Gordon 1987
}
