\name{montpellier}
\alias{montpellier}
\docType{data}
\title{Echantillons du puit de la rue Barralerie.

}
\description{
Jeu de données issu de Waksman et al. 1999. Echantillons du puit de la rue Barralerie, 1999
}
\usage{data("montpellier")}
\format{
  A data frame with 100 observations on the following 25 variables.
  \describe{
    \item{\code{nom}}{a factor with levels \code{NMA479} \code{NMA480} \code{NMA481} \code{NMA482} \code{NMA512} \code{NMA513} \code{NMA514} \code{NMA515} \code{NMA516} \code{NMA517} \code{NMA518} \code{NMA530} \code{NMA531} \code{NMA532} \code{NMA533} \code{NMA534} \code{NMA535} \code{NMA536} \code{NMA537} \code{NMA538} \code{NMA539} \code{NMA540} \code{NMA541} \code{NMA613} \code{NMA614} \code{NMA615} \code{NMA616} \code{NMA617} \code{NMA618} \code{NMA619} \code{NMA620} \code{NMA621} \code{NMA622} \code{NMA623} \code{NMA624} \code{NMA625} \code{NMA626} \code{NMA627} \code{NMA628} \code{NMA629} \code{NMA690} \code{NMA691} \code{NMA692} \code{NMA693} \code{NMA694} \code{NMA695} \code{NMA696} \code{NMA697} \code{NMA698} \code{NMA699} \code{NMA784} \code{NMA785} \code{NMA786} \code{NMA787} \code{NMA788} \code{NMA789} \code{NMA790} \code{NMA791} \code{NMA792} \code{NMA793} \code{NMA794} \code{NMA795} \code{TWM  1} \code{TWM  2} \code{TWM  3} \code{TWM  4} \code{TWM  5} \code{TWM  6} \code{TWM  7} \code{TWM  8} \code{TWM  9} \code{TWM 10} \code{TWM 11} \code{TWM 12} \code{TWM 13} \code{TWM 14} \code{TWM 15} \code{TWM 16} \code{TWM 17} \code{TWM 18} \code{TWM 19} \code{TWM 20} \code{TWM 21} \code{TWM 22} \code{TWM 23} \code{TWM 24} \code{TWM 25} \code{TWM 26} \code{TWM 27} \code{TWM 28} \code{TWM 29} \code{TWM 30} \code{TWM 54} \code{TWM 55} \code{TWM 56} \code{TWM 57} \code{TWM 58} \code{TWM 59} \code{TWM 60} \code{TWM 61}}
    \item{\code{CaO}}{a numeric vector}
    \item{\code{Fe2O3}}{a numeric vector}
    \item{\code{TiO2}}{a numeric vector}
    \item{\code{K2O}}{a numeric vector}
    \item{\code{SiO2}}{a numeric vector}
    \item{\code{Al2O3}}{a numeric vector}
    \item{\code{MgO}}{a numeric vector}
    \item{\code{MnO}}{a numeric vector}
    \item{\code{Na2O}}{a numeric vector}
    \item{\code{P2O5}}{a numeric vector}
    \item{\code{Zr}}{a numeric vector}
    \item{\code{Sr}}{a numeric vector}
    \item{\code{Rb}}{a numeric vector}
    \item{\code{Zn}}{a numeric vector}
    \item{\code{Cr}}{a numeric vector}
    \item{\code{Ni}}{a numeric vector}
    \item{\code{La}}{a numeric vector}
    \item{\code{Ba}}{a numeric vector}
    \item{\code{V}}{a numeric vector}
    \item{\code{Ce}}{a numeric vector}
    \item{\code{Y}}{a numeric vector}
    \item{\code{Th}}{a numeric vector}
    \item{\code{Pb}}{a numeric vector}
    \item{\code{Cu}}{a numeric vector}
  }
}

\source{
Montpellier, Terre de faïence, SilvanaEditorale 2010
}

\examples{
data(montpellier)
## maybe str(montpellier) ; plot(montpellier) ...
}
\keyword{datasets}
