% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travail_sur_groupe.R
\name{travail_sur_groupe}
\alias{travail_sur_groupe}
\title{travail sur groupe}
\usage{
travail_sur_groupe(data, pas, normalize = TRUE, echantillonage = TRUE)
}
\arguments{
\item{data}{dataset de base}

\item{pas}{pas}

\item{normalize}{permet de ramener à 100}

\item{echantillonage}{afficher ou non le nom des individus}
}
\value{
deux histogrammes
}
\description{
La distance de Mahalanobis permet d'identifier
}
