% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_discriminante.R
\name{ana_dis}
\alias{ana_dis}
\title{analyse discriminante}
\usage{
ana_dis(data, pas, echantillon, normalize = TRUE, echantillonage = TRUE)
}
\arguments{
\item{data}{dataset de base}

\item{pas}{pas}

\item{echantillon}{dataset a analyser}

\item{normalize}{permet de ramener à 100}

\item{echantillonage}{vrai ou faux}
}
\value{
deux histogrammes
}
\description{
La distance de Mahalanobis permet d'identifier
}
