% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_de_grappe.R
\name{grappe}
\alias{grappe}
\title{Analyse de grappe}
\usage{
grappe(
  data,
  normalize = TRUE,
  methode = "simple",
  h = NULL,
  k = NULL,
  color = FALSE
)
}
\arguments{
\item{data}{Un dataframe.}

\item{normalize}{Un booleen. Si TRUE, les datas sont ramenées à 100 avant la classification, sinon elles sont gardées telles quelles.}

\item{methode}{La methode de transformation utilisée au préalable de la classification. Peut être \code{"simple"} ou \code{"logarithme"}.}

\item{k, h}{Valeur scalaire. Coupe le dendogramme pour former \code{k} clusters ou à une hauteur \code{h}.}

\item{color}{Un booleen. Si TRUE, le nom de chaque individu est colorié selon son groupe de référence. La dernière colonne de \code{data} doit alors être le numéro du groupe de référence.}
}
\value{
Retourne une liste contenant:
\item{data}{le fichier data triée dans l'ordre du dendogramme. Si \code{k} ou \code{h} est renseigné, une colonne \code{Groupe} est automatiquement ajoutée avec le numéro du groupe auquel appartient chaque échantillon}
\item{split}{une liste avec les éléments du datasets pour chaque groupe}
\item{hc}{l'arbre hiérarchique est retourner sous forme d'un objet de classe hclust. (Voir fonction R \code{hclust} pour plus de détails)}
\item{descriptif}{Si \code{k} ou \code{h} est renseigné. Retourne la liste contenant les tests statistiques réalisés sur les groupes.}
}
\description{
L'analyse de grappe ou classification hiérarchique ascendante est une technique de classification automatique en analyse de données.
Contrairement à une analyse factorielle, elle prend en compte toutes les dimensions d'un problème, sans perte d'informations. Mais son principal avantage réside dans l'interprétation facile du résultat.
}
\details{
La fonction affiche un arbre hiérarchique.
Un ensemble de tests statistiques sont appliqués aux valeurs groupées pour vérifier la pertinence des groupes. On obtient une liste des éléments caractéristiques pour chaque groupe, avec leur valeur-test. Un test d'identitié des moyennes est également appliqué pour l'ensemble des variables.\cr
\cr
Le calcul préalable de traitement des données par la méthode \code{"logarithme"} s'appuie la transformation centrée log-ratio introduite par Aitchison en 1986. Il permet d'éviter le phénomène de contamination de la matrice des covariances en supprimant une contrainte.

\cr
L'arbre est construit avec une matrice des distances euclidiennes et par affinité moyenne, conformément à l'usage en archéométrie.
}
\examples{

## Not run:
## Données de Yokman 2010
data('rock')

## Analyse de grappe
#grappe(rock)

# Les éléments 36, 38 et 42 sont marginaux, on les exclus
rock<-exclure(rock,'46')
rock<-exclure(rock,'38')
rock<-exclure(rock,'42')

## Deuxième analyse
#grappe(rock)

### Identification de quatre groupes
#grappe(rock,k=4)

## End(Not run)

## Exemple 2 Waksman, 1999


}
