% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_de_grappe.R
\name{grappe}
\alias{grappe}
\title{Analyse de grappe}
\usage{
grappe(data, normalize = TRUE, methode = "simple", h = NULL, k = NULL)
}
\arguments{
\item{data}{Un dataframe.}

\item{normalize}{Un booleen. Si TRUE, les datas sont ramenées à 100 avant la classification, sinon elles sont gardées telles quelles.}

\item{methode}{La methode de transformation utilisée au préalable de la classification. Peut être \code{"simple"} ou \code{"logarithme"}.}

\item{k, h}{Valeur scalaire. Coupe le dendogramme pour former \code{k} clusters ou à une hauteur \code{h}.}
}
\value{
Une liste de la classe hclust.
}
\description{
L'analyse de grappe ou classification hiérarchique ascendante est une technique de classification automatique en analyse de données.
Contrairement à une analyse factorielle, elle prend en compte toutes les dimensions d'un problème, sans perte d'informations. Mais son principal avantage réside dans l'interprétation facile du résultat.
}
\details{
La fonction affiche un arbre hiérarchique. Une liste dataframe \code{data_triee} est automatiquement créée avec les échantillons rangées dans leur ordre d'apparition dans l'arbre.  Si le paramètre \code{k} ou \code{h} est renseigné, alors les groupes formés apparaissent sur le graphique et une colonne \code{Groupe} est ajoutée au fichier \code{data_triee}.
Un ensemble de tests statistiques sont appliqués aux valeurs groupées pour vérifier la pertinence des groupes. On obtient une liste des éléments caractéristiques pour chaque groupe, avec leur valeur-test. Un test d'identitié des moyennes est également appliqué pour l'ensemble des variables.\cr
\cr
Le calcul préalable de traitement des données par la méthode \code{"logarithme"} s'appuie la transformation centrée log-ratio introduite par Aitchison en 1986. Il permet d'éviter le phénomène de contamination de la matrice des covariances en supprimant une contrainte.
}
\examples{

## Not run:
## Données de Yokman 2010
data('rock')

## Analyse de grappe
#grappe(rock)

# Les éléments 36, 38 et 42 sont marginaux, on les exclus
rock<-exclure(rock,'46')
rock<-exclure(rock,'38')
rock<-exclure(rock,'42')

## Deuxième analyse
#grappe(rock)

### Identification de quatre groupes
#grappe(rock,k=4)

## End(Not run)

## Exemple 2 Waksman, 1999


}
