% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travail_sur_groupe.R
\name{travail_sur_groupe}
\alias{travail_sur_groupe}
\title{Travail sur groupe}
\usage{
travail_sur_groupe(data, pas, normalize = TRUE, echantillonage = TRUE)
}
\arguments{
\item{data}{Un dataframe.}

\item{pas}{Un sclaire. Il correspond à la largeur des intervalles de histogramme.}

\item{normalize}{Un booleen. Si TRUE, les datas sont ramenées à 100 avant la classification, sinon elles sont gardées telles quelles.}

\item{echantillonage}{Un booleen. Si TRUE, le nom des échantillons apparaît sur l'histogramme.}
}
\value{
deux histogrammes
}
\description{
La fonction \code{travail_sur_groupe()} permet de tracer l'histogramme des distances de Mahalannobis au sein du groupe.
Elle reprend une fonctionnalité disponible sur le site de ceramologie. Pour plus d'information sur les distances de Mahalanobis, voir la documentation \code{ana_dis}.
}
\examples{
#Plus tard

}
